tools/requirements.txt
pyyaml
bs4
Pillow
